requirements.txt
django >=4.2,<4.3
django-cors-headers
django-mptt
djangorestframework
psycopg2 >=2.9,<3.0

langchain
openai
python_dotenv

# development only
black